avoid a variety of warnings for MSVC
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@21684 f0d54148-0727-0410-94bb-9a71ac55c965

+++ ogr/ogrsf_frmts/bna/ogrbnalayer.cpp
@@ -169,7 +169,7 @@ OGRFeature *OGRBNALayer::GetNextFeature()
     while(1)
     {
         int ok = FALSE;
-        offset = VSIFTellL(fpBNA);
+        offset = (int) VSIFTellL(fpBNA);
         line = curLine;
         if (nNextFID < nFeatures)
         {
@@ -787,7 +787,7 @@ void OGRBNALayer::FastParseUntil ( int interestFID)
         while(1)
         {
             int ok = FALSE;
-            int offset = VSIFTellL(fpBNA);
+            int offset = (int) VSIFTellL(fpBNA);
             int line = curLine;
             record =  BNA_GetNextRecord(fpBNA, &ok, &curLine, TRUE, BNA_READ_NONE);
             if (ok == FALSE)
